# Conflicts:
#	src/main/java/com/sw19/sofa/domain/article/service/ArticleService.java
#	src/main/java/com/sw19/sofa/domain/linkcard/service/LinkCardService.java
#	src/main/java/com/sw19/sofa/domain/linkcard/service/LinkCardTagService.java
#	src/main/java/com/sw19/sofa/domain/tag/service/CustomTagService.java
#	src/main/java/com/sw19/sofa/domain/tag/service/TagService.java

+++ src/main/java/com/sw19/sofa/domain/remind/service/RemindService.java
@@ -6,6 +6,7 @@ import com.sw19.sofa.domain.remind.entity.Remind;
 import com.sw19.sofa.domain.remind.enums.RemindSortBy;
 import com.sw19.sofa.domain.remind.repository.RemindRepository;
 import com.sw19.sofa.global.common.dto.enums.SortOrder;
+import com.sw19.sofa.global.scheduler.manager.SchedulerManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,6 +18,23 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class RemindService {
     private final RemindRepository remindRepository;
+    private final SchedulerManager schedulerManager;
+
+    @Transactional
+    public void addAllByLinkCardListAndMember(List<LinkCard> linkCardList, Member member){
+        List<Remind> remindList = linkCardList.stream().map(linkCard ->
+                Remind.builder()
+                        .linkCard(linkCard)
+                        .member(member)
+                        .build()
+        ).toList();
+
+        remindRepository.saveAll(remindList);
+    }
+
+    public List<Remind> getRemindListByMember(Member member){
+        return remindRepository.findAllByMember(member);
+    }
 
     public List<Remind> searchReminds(
             Long memberId,
@@ -29,11 +47,13 @@ public class RemindService {
     }
 
     @Transactional
-    public void deleteRemind(LinkCard linkCard, Member member) {
-        remindRepository.deleteByLinkCardAndMember(linkCard, member);
-    }
+    public void removeFromRemind(LinkCard linkCard, Member member) {
 
-    public boolean existsRemind(LinkCard linkCard, Member member) {
-        return remindRepository.existsByLinkCardAndMember(linkCard, member);
+        if (remindRepository.existsByLinkCardAndMember(linkCard, member)) {
+            remindRepository.deleteByLinkCardAndMember(linkCard, member);
+            if(!remindRepository.existsByMember(member)){
+                schedulerManager.stopMemberNotifyScheduler(member.getEncryptUserId());
+            }
+        }
     }
 }
